backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.1
requests==2.31.0

